add viz notebook + plot refactoring
